+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+## 0.0.2 (2017-01-20, alpha release)
+
+#### features:
+ - added PPSP.snap(), a function to snap to the closest target DOM
+ - added PPSP.refresh(), a function to refresh and reload the selector DOMs
+ - added mouse wheel navigation
+ - added adjusting position on resize
+
 ## 0.0.1 (2017-01-20, alpha release)
 
 #### features: